Split up `decode_test` into tests focused on individual SQL protos.
In PR #421, @bgogul noted that `decode_test.cc` was rather long and had
little code sharing between its tests and asked that it be split up to
improve readability. This PR does that.
